requirements.txt

Django==4.1.7
django-restframework==0.0.1

djangorestframework-simplejwt==4.8.0
djoser==2.1.0

pymysql==1.0.2

PyJWT==2.6.0


dj-rest-auth==2.2.5

django-allauth==0.51.0

drf-yasg==1.21.4

Pillow==9.3

pandas==1.5.3

openpyxl==3.1.1

